backend/requirements.txt
Flask==3.0.2
boto3==1.26.137
python-dotenv==0.19.0
Pillow==10.2.0
flask-cors==3.0.10
python-jose==3.3.0
gunicorn==20.1.0 